+++ .gitignore
@@ -1,0 +1,9 @@
+## Ignore Visual Studio temporary files, build results, and
+## files generated by popular Visual Studio add-ons.
+
+*.opendb
+*.db
+*.suo
+int/
+packages/
+.vs/
