+++ Lab2/Lab2.vcxproj
@@ -144,12 +144,14 @@
     <ClCompile Include="fileHandler.c" />
     <ClCompile Include="Lab5.c" />
     <ClCompile Include="off.c" />
+    <ClCompile Include="main.c" />
     <ClCompile Include="userInput.c" />
     <ClCompile Include="vector.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="example.h" />
     <ClInclude Include="fileHandler.h" />
+    <ClInclude Include="Lab5.h" />
     <ClInclude Include="off.h" />
     <ClInclude Include="userInput.h" />
     <ClInclude Include="vector.h" />
