+++ Lab2/Lab2.vcxproj.filters
@@ -33,6 +33,9 @@
     <ClCompile Include="userInput.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="main.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="example.h">
@@ -48,6 +51,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="vector.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Lab5.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
