+++ Lab2/Lab5.c
@@ -1,6 +1,4 @@
-#include <stdio.h>
 #include "off.h"
-#include <gl/freeglut.h>
 
 // for mouse things
 float deltaAngle = 0.0f;
@@ -14,8 +12,6 @@ static float viewer[] = {
 enum {
 	ORTHO2D, PERSPECTIVE, FRUSTUM, ORTHO
 } mode = ORTHO;
-
-Model model;
 
 void drawOrigin() {
 	point3 origin[4] =
@@ -161,22 +157,4 @@ void mouseMove(int x, int y) {
 	viewer[3] = sin(deltaAngle);
 	viewer[5] = -cos(deltaAngle);
 	glutPostRedisplay();
-}
-
-int main(int argc, char** argv)
-{
-	const char* fileName = "bone_normalized_aligned.off";
-	model = readOFFFile(fileName);
-	glutInit(&argc, argv);
-	glutInitDisplayMode(GLUT_SINGLE | GLUT_RGB | GLUT_DEPTH);
-	glutInitWindowSize(500, 500);
-	glutInitWindowPosition(110, 0);
-	glutCreateWindow("Hello OpenGL");
-	glutPassiveMotionFunc(mouseMove);
-	glutKeyboardFunc(keys);
-	glutSetCursor(GLUT_CURSOR_CROSSHAIR);
-	glutDisplayFunc(mydisplay);
-	init();
-	glutMainLoop();
-	return 0;
 }
