+++ Lab2/Lab5.h
@@ -1,0 +1,12 @@
+#pragma once
+#include "off.h"
+
+void drawOrigin();
+
+void init(void);
+
+void mydisplay(void);
+
+void keys(unsigned char key, int x, int y);
+
+void mouseMove(int x, int y);
