+++ Lab2/x64/Debug/Lab5.exe.recipe
@@ -2,7 +2,7 @@
 <Project>
   <ProjectOutputs>
     <ProjectOutput>
-      <FullPath>C:\UNI\Current\ICT289 - Computer Graphics Principles and Programming\Lab5_ex3\x64\Debug\Lab5.exe</FullPath>
+      <FullPath>C:\Repositories\c-world\x64\Debug\Lab5.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
