+++ Lab2/x64/Debug/Lab5.tlog/Lab5.lastbuildstate
@@ -1,2 +1,2 @@
 PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.29.30037:VCServicingVersionATL=14.29.30038:VCServicingVersionCrtHeaders=14.29.30038:VCServicingVersionCompilers=14.29.30038:TargetPlatformVersion=10.0.19041.0:
-Debug|x64|C:\UNI\Current\ICT289 - Computer Graphics Principles and Programming\Lab5_ex3\|
+Debug|x64|C:\Repositories\c-world\|
